Cleans up add to storage
