+++ src/UpdateStorage.ts
@@ -6,44 +6,36 @@ import {
 } from './common/HelperFunctions';
 import { extensionStateKey } from './common/Constants';
 
-export function listenerForAddToStorage(
+export async function addToStorage(
   dataForPath: Map<string, Map<number, Record<string, unknown>>>,
-  currentPathName: string): void {
+  currentPathName: string,
+  newData: Record<string, string>): Promise<void> {
 
-  const queryBtn: HTMLElement = document.querySelector(`input[name='${currentPathName}Submit']`);
-  const textarea: HTMLTextAreaElement = document.querySelector('textarea');
-
-  queryBtn.onclick = async () => {
-    const data = textarea.value.trim();
-    if (!data) {
-      return; // If textarea is empty, do nothing
-    }
-    const { formattedDate, timestamp: newTimestamp } = getFormattedDateAndTimestamp();
-    const newData = { data };
-    const dataForPathArray = [...dataForPath];
-    /* 
-      Since we can only insert at TODAY's date (and not in past or future), that means we only need to check 
-      the latest date if it exists. If the date exists, that means we need to update that date's map. 
-      If it doesn't exist, that means we need a new entry for that date
-    */
-    const latestDate = dataForPathArray.length > 0 && dataForPathArray[0][0];
-    if (latestDate && (latestDate === formattedDate)) {
-      // convert the existing details map to array to insert into beginning
-      const existingMapArray = [...dataForPathArray[0][1]];
-      existingMapArray.unshift([newTimestamp, newData]); // insert the new record
-      dataForPathArray[0][1] = new Map(existingMapArray); // convert back to map
-    } else {
-      const newMap = new Map<number, Record<string, unknown>>();
-      newMap.set(newTimestamp, newData);
-      dataForPathArray.unshift([formattedDate, newMap]); // insert new record in the data
-    }
-    const dateMap = new Map<string, string>();
-    for (const [date, info] of dataForPathArray) { // Serialize data again before storing
-      dateMap.set(date, serializeMap<number, Record<string, unknown>>(info));
-    }
-
-    await setDataToPath(currentPathName, serializeMap<string, string>(dateMap));
+  const { formattedDate, timestamp: newTimestamp } = getFormattedDateAndTimestamp();
+  const dataForPathArray = [...dataForPath];
+  /* 
+    Since we can only insert at TODAY's date (and not in past or future), that means we only need to check 
+    the latest date if it exists. If the date exists, that means we need to update that date's map. 
+    If it doesn't exist, that means we need a new entry for that date
+  */
+  const latestDate = !!dataForPathArray.length && dataForPathArray[0][0];
+  if (latestDate && (latestDate === formattedDate)) {
+    // convert the existing details map to array to insert into beginning
+    const existingMapArray = [...dataForPathArray[0][1]];
+    existingMapArray.unshift([newTimestamp, newData]); // insert the new record
+    dataForPathArray[0][1] = new Map(existingMapArray); // convert back to map
+  } else {
+    const newMap = new Map<number, Record<string, unknown>>();
+    newMap.set(newTimestamp, newData);
+    dataForPathArray.unshift([formattedDate, newMap]); // insert new record in the data
   }
+  const dateMap = new Map<string, string>();
+  for (const [date, info] of dataForPathArray) { // Serialize data again before storing
+    dateMap.set(date, serializeMap<number, Record<string, unknown>>(info));
+  }
+
+  await setDataToPath(currentPathName, serializeMap<string, string>(dateMap));
+
 }
 
 export async function updateExtensionState(currentPathName: string, newState: boolean): Promise<void> {
